backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0
python-multipart>=0.0.12
sqlalchemy>=2.0.0
alembic>=1.13.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
httpx>=0.24.0
click>=8.1.0
cryptography>=41.0.0
aiofiles>=23.2.0
python-slugify>=8.0.0
pillow>=10.0.0
aiosmtplib>=3.0.0
pandas>=2.0.0
openpyxl>=3.1.0
